requirements.txt
numpy~=1.26.3
opencv-python~=4.9.0.80
fastapi~=0.108.0
pillow~=10.2.0
tensorflow~=2.15.0
uvicorn
python-multipart
tensorflow_hub